+++ deps.ts
@@ -9,11 +9,22 @@ export type {
   RouteContext,
 } from "jsr:@oak/oak@^16.1.0";
 
+import {
+  type ResponseConfig,
+  type RouteConfig,
+} from "npm:@asteasolutions/zod-to-openapi@^7.1.1";
+
+export type OakOpenApiSpec = Omit<RouteConfig, "method" | "path"> & {
+  responses?: {
+    [statusCode: string]: ResponseConfig;
+  };
+};
+
+export { type ResponseConfig, type RouteConfig };
+
 export {
   OpenApiGeneratorV3,
   OpenAPIRegistry,
-  type ResponseConfig,
-  type RouteConfig as OasRouteConfig,
   type ZodRequestBody,
 } from "npm:@asteasolutions/zod-to-openapi@^7.1.1";